backend/requirements.txt
Flask
numpy
Pillow
tensorflow
opencv-python
react-router-dom
react react-dom
jspdf html2canvas
firebase-admin
react-icons
axios